+++ src/onegov.file/onegov/file/__init__.py
@@ -1,13 +1,20 @@
 from onegov.file.collection import FileCollection, FileSetCollection
 from onegov.file.integration import DepotApp
-from onegov.file.models import File, FileSet, AssociatedFiles, SearchableFile
+from onegov.file.models import (
+    File,
+    FileMessage,
+    FileSet,
+    AssociatedFiles,
+    SearchableFile
+)
 
-__all__ = [
+__all__ = (
     'AssociatedFiles',
     'DepotApp',
     'File',
     'FileCollection',
+    'FileMessage',
     'FileSet',
     'FileSetCollection',
     'SearchableFile',
-]
+)