+++ src/onegov.file/onegov/file/integration.py
@@ -19,7 +19,7 @@ from onegov.file.collection import FileCollection
 from onegov.file.errors import AlreadySignedError
 from onegov.file.errors import InvalidTokenError
 from onegov.file.errors import TokenConfigurationError
-from onegov.file.models import File
+from onegov.file.models import File, FileMessage
 from onegov.file.sign import SigningService
 from onegov.file.utils import digest, current_dir
 from pathlib import Path
@@ -278,6 +278,9 @@ class DepotApp(App):
         }
 
         file.signed = True
+
+        FileMessage.log_signature(file, signee)
+
         session.flush()
 
     @property
@@ -478,4 +481,8 @@ def delete_file(self, request):
 
     """
     request.assert_valid_csrf_token()
+
+    if self.signed:
+        FileMessage.log_signed_file_removal(self, request.current_username)
+
     FileCollection(request.session).delete(self)