+++ src/onegov.file/onegov/file/models/__init__.py
@@ -1,5 +1,12 @@
 from onegov.file.models.file import File, SearchableFile
 from onegov.file.models.fileset import FileSet
 from onegov.file.models.associated_files import AssociatedFiles
+from onegov.file.models.file_message import FileMessage
 
-__all__ = ['AssociatedFiles', 'File', 'FileSet', 'SearchableFile']
+__all__ = (
+    'AssociatedFiles',
+    'File',
+    'FileMessage',
+    'FileSet',
+    'SearchableFile'
+)