+++ src/onegov.file/HISTORY.rst
@@ -1,6 +1,9 @@
 Changelog
 ---------
 
+- Adds log messages when signing a file and when removing a signed file.
+  [href]
+
 0.11.0 (2018-10-09)
 ~~~~~~~~~~~~~~~~~~~